+++ src/DataFactory/DataFactoryV2/Constants.cs
@@ -260,6 +260,10 @@ namespace Microsoft.Azure.Commands.DataFactoryV2
 
         public const string HelpGlobalParameter = "The dictionary containing the global parameters of the data factory.";
 
+        public const string HelpIdentityType = "The identity type for the factory. It can be 'SystemAssigned', 'UserAssigned' or 'SystemAssigned,UserAssigned'";
+
+        public const string HelpUserAssignedIdentities = "Dictionary of user assigned identities to use. { \"<UAMI Id>\" : {} }";
+
         public const string HelpRepoConfigurationProjectName = "The project name for repo configuration.";
 
         public const string HelpRepoConfigurationTenantId = "The tenant id for repo configuration.";