+++ src/Resources/Resources/Properties/Resources.Designer.cs
@@ -484,11 +484,29 @@ namespace Microsoft.Azure.Commands.Resources.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to No AD object was found with the parameters provided please ensure that the display name or GUID is written properly.
+        /// </summary>
+        internal static string NoADObjectFound {
+            get {
+                return ResourceManager.GetString("NoADObjectFound", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Cannot find any deployments in resource group &apos;{0}&apos;..
         /// </summary>
         internal static string NoDeploymentWereFound {
             get {
                 return ResourceManager.GetString("NoDeploymentWereFound", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Call to Graph was unsuccesfull, this is likely due to insuficient permissions in Azure AD. Please ensure you have the correct set of permissions..
+        /// </summary>
+        internal static string NotAuthorizedInGraph {
+            get {
+                return ResourceManager.GetString("NotAuthorizedInGraph", resourceCulture);
             }
         }