+++ src/Storage/Storage.Management.Test/ScenarioTests/StorageAccountTests.cs
@@ -225,5 +225,19 @@ namespace Microsoft.Azure.Commands.Management.Storage.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-NewSetAzureStorageAccountAllowSharedKeyAccess");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestNewAzureStorageAccountEdgeZone()
+        {
+            TestRunner.RunTestScript("Test-NewAzureStorageAccountEdgeZone");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestAzureStorageAccountKeySASPolicy()
+        {
+            TestRunner.RunTestScript("Test-AzureStorageAccountKeySASPolicy");
+        }
     }
 }