+++ src/Storage/Storage.Management.Test/ScenarioTests/StorageAccountTests.cs
@@ -239,5 +239,12 @@ namespace Microsoft.Azure.Commands.Management.Storage.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-AzureStorageAccountKeySASPolicy");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestAzureStorageAccountUserAssignedIdentity()
+        {
+            TestRunner.RunTestScript("Test-AzureStorageAccountUserAssignedIdentity");
+        }        
     }
 }