+++ src/DataFactory/DataFactoryV2/Changelog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added User Assigned Identities to Data Factory.
 
 ## Version 1.11.5
 * Updated ADF .Net SDK version to 4.15.0