+++ src/DataFactory/DataFactoryV2/Changelog.md
@@ -18,7 +18,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
 * Added User Assigned Identities to Data Factory.
+* Updated ADF .Net SDK version to 4.17.1
 
 ## Version 1.11.5
 * Updated ADF .Net SDK version to 4.15.0