+++ src/DataFactory/DataFactoryV2/Changelog.md
@@ -20,6 +20,7 @@
 ## Upcoming Release
 * Added User Assigned Identities to Data Factory.
 * Updated ADF .Net SDK version to 4.17.1
+* Updated ADF encryption client SDK version to 5.5.7773.2
 
 ## Version 1.11.5
 * Updated ADF .Net SDK version to 4.15.0